Moved error handling functions into own file
